+++ src/main/java/bridge/Application.java
@@ -3,6 +3,6 @@ package bridge;
 public class Application {
 
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+        BridgeGame.getBridgeGame().runGame();
     }
 }
